Convert to i64

+++ src/common_ops.rs
@@ -1,8 +1,8 @@
-pub fn get_numbers(number_list_string: &str) -> Vec<i32> {
+pub fn get_numbers(number_list_string: &str) -> Vec<i64> {
     let number_strings = number_list_string.split_ascii_whitespace();
     let mut numbers = Vec::new();
     for number_str in number_strings {
-        let number: i32 = number_str.parse().unwrap();
+        let number: i64 = number_str.parse().unwrap();
         numbers.push(number);
     }
     numbers
